// src/app/shared/model/labtestreport.ts
export class Labtestreport {
    LtreportId: number = 0;
    AppointmentId: number = 0; 
    TpId: number = 0;
    LabTestId: number = 0; 
    ActualResult: string = "";
    Remarks: string = "";
    TestName: string = "";
    PatientName: string = "";
    LowRange: number = 0; 
    HighRange: number = 0;
    SampleItem: string = "";
}